removed older config and added locality label changes
