+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/Project/src/pages/employee/index.js
@@ -2,7 +2,6 @@ import React, { useEffect } from "react";
 import { Switch, useLocation } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { PrivateRoute, AppContainer, BreadCrumb } from "@egovernments/digit-ui-react-components";
-import CreateProjectMUKTA from "./CreateProjectMUKTA";
 import WorkflowCompTest from "./WorkflowCompTest";
 
 const ProjectBreadCrumb = ({ location }) => {
@@ -68,7 +67,7 @@ const App = ({ path }) => {
         <React.Fragment>
           <ProjectBreadCrumb location={location} />
         </React.Fragment>
-        <PrivateRoute path={`${path}/create-project`} component={() => <CreateProjectMUKTA parentRoute={path}/>} />
+        <PrivateRoute path={`${path}/create-project`} component={() => <CreateProjectComponent parentRoute={path}/>} />
         <PrivateRoute path={`${path}/project-details`} component={() => <ProjectDetailsComponent parentRoute={path}/>} />
         <PrivateRoute path={`${path}/search-project`} component={() => <ProjectSearchComponent parentRoute={path}/>} />
         <PrivateRoute path={`${path}/inbox`} component={() => <ProjectSearchAndInboxComponent parentRoute={path}/>} />
